Refactor application layout on windows
Now only the calibre executables are present in the top level directory
(which is added to PATH). No third party executables and no dlls.

This became necessary because the VS 2015 CRT is not manifest based, so
it would have to be placed in the top level dir in the old layout. Now
it can be put into the DLLs dir.